refactor(LottoNumbers): LottoNumbers 가 6개의 로또 숫자를 가지고 행하는 역할을 하도록 수정
- Lotto -> LottoNumbers 네이밍 변경
- LottoVendingMachine 에서 랜덤 6자리 수를 생성하도록 추가
