+++ src/main/kotlin/lotto/domain/WinningStatistics.kt
@@ -25,8 +25,8 @@ data class WinningStatistics(
         private val INITIAL_STATISTICS = Rank.values().associateWith { INITIAL_COUNT }
         private val RATE_OF_RETURN_ROUNDING_MODE = RoundingMode.FLOOR
 
-        fun create(lottoNumbersList: List<Lotto>, winningLottoNumbers: WinningLotto): WinningStatistics {
-            val matchResult = winningLottoNumbers.getMatchResult(lottoNumbersList)
+        fun create(lottoNumbersNumbersList: List<LottoNumbers>, winningLottoNumbers: WinningLotto): WinningStatistics {
+            val matchResult = winningLottoNumbers.getMatchResult(lottoNumbersNumbersList)
             val ranks = matchResult.map { Rank.valueOf(it.first, it.second) }
 
             return WinningStatistics(statistics = makeStatistics(ranks))
