+++ src/main/kotlin/lotto/view/ResultView.kt
@@ -1,6 +1,6 @@
 package lotto.view
 
-import lotto.domain.Lotto
+import lotto.domain.LottoNumbers
 import lotto.domain.Rank
 import lotto.domain.WinningStatistics
 import java.math.BigDecimal
@@ -27,7 +27,7 @@ object ResultView {
         if (rateOfReturn < BigDecimal(1)) println("(기준이 1이기 때문에 결과적으로 손해라는 의미임)")
     }
 
-    fun printPurchasedLotto(myLotteries: List<Lotto>) {
+    fun printPurchasedLotto(myLotteries: List<LottoNumbers>) {
         myLotteries.forEach { lotto ->
             val sortedValue = lotto.lottoNumbers.sortedWith(compareBy { it.number })
             println("[${sortedValue.joinToString(", ")}]")
